Added support for per argument help using docstring in commands Moved Room, Exit from world to room module
