+++ abacura-core/abacura/plugins/commands/__init__.py
@@ -125,27 +125,45 @@ class Command:
     def get_help(self):
         help_text = []
 
-        doc = getattr(self.callback, '__doc__', None)
+        doc_lines = []
+
+        parameter_doc = {}
+        for line in getattr(self.callback, '__doc__', '').split("\n"):
+            if line.strip().startswith(":") and line.strip().find(" ") >= 0:
+                s = line.lstrip(" :").split(" ")
+                name = s[0]
+                description = " ".join(s[1:])
+                parameter_doc[name] = description
+            elif len(line.strip(" \n")):
+                doc_lines.append(line)
 
         parameters = self.get_parameters()
+        parameter_names = []
         parameter_help = []
         for parameter in parameters:
             if parameter.default is inspect.Parameter.empty:
-                parameter_help.append(parameter.name)
+                parameter_names.append(parameter.name)
             else:
-                parameter_help.append(f"[{parameter.name}]")
+                parameter_names.append(f"[{parameter.name}]")
+            if parameter.name in parameter_doc:
+                parameter_help.append(f"    {parameter.name:30s} : {parameter_doc.get(parameter.name)}")
 
-        if doc is not None:
-            help_text.append(doc + "\n")
+        if len(doc_lines):
+            help_text.append("\n".join(doc_lines) + "\n")
 
-        help_text.append(f"  Usage: {self.name} {' '.join(parameter_help)}")
+        help_text.append(f"  Usage: {self.name} {' '.join(parameter_names)}")
+
+        if len(parameter_help):
+            help_text.append("")
+            help_text.append("\n".join(parameter_help))
 
         option_help = []
         for name, p in self.get_options().items():
             if p.annotation in (bool, 'bool'):
-                option_help.append(f"  --{name.lstrip('_')}")
+                option_help.append(f"  --{name.lstrip('_'):30s} : {parameter_doc.get(name, '')}")
             else:
-                option_help.append(f"  --{name.lstrip('_')}=<{p.annotation.__name__}> ")
+                pname = f"{name.lstrip('_') + '=<' + p.annotation.__name__ + '>'}"
+                option_help.append(f"  --{pname:30s} : {parameter_doc.get(name, '')}")
 
         if len(option_help):
             help_text.append("\nOptions:\n")
